Fix Snap Peak popup crash
